fix: middleware -> proxy
